Fix int overflow error
The validate check at line 76 doesn't work for N values above 50K or so because when you square it, the value overflows and winds up negative (i.e. N * (N - 1) / 2 can end up lower than M and so appears invalid). The N value should be parsed as a Long on line 74 to avoid the overflow error.